backend/requirements.txt
fastapi>=0.95.2
uvicorn>=0.22.0
python-multipart>=0.0.18
scikit-learn>=1.3.0
nltk>=3.8.1
PyMuPDF>=1.22.5
python-docx>=0.8.11
numpy>=1.23.5
